fixed bitcoin stoping on yetiwarm

+++ appyetiwarm.py
@@ -49,6 +49,7 @@ adrlist = []
 transnum = 0
 progress = 0
 samedesc = False
+sent = False
 utxo = None
 rescan = False
 switcher = {
@@ -198,12 +199,17 @@ def BTCFinished():
         bitcoinprogress = True
     return bitcoinprogress
 
-def BTCRunning():
+def BTCClosed():
     home = os.getenv("HOME")
     if (subprocess.call('lsof -n -i :8332', shell=True) != 1):
-        return True
+        return False
     elif os.path.exists(home + "/.bitcoin/bitcoind.pid"):
         subprocess.call('rm -r ~/.bitcoin/bitcoind.pid', shell=True)
+    return True
+
+def BTCRunning():
+    if not (BTCprogress() == 0):
+        return True
     return False
 
 def RPCYW():
@@ -324,17 +330,17 @@ def YWrestartbitcoin():
     global progress
     global IBD
     if request.method == 'GET':
-        subprocess.call('python3 ~/yeticold/utils/stopbitcoin.py', shell=True)
-        subprocess.call('rm -r ~/.bitcoin/yetiwarm*', shell=True)
-        subprocess.call('rm -r ~/yetiwarmwallet*', shell=True)
-        subprocess.Popen('~/yeticold/bitcoin-0.19.0rc1/bin/bitcoin-qt -proxy=127.0.0.1:9050',shell=True,start_new_session=True)
-        progress = BTCprogress()
+        if BTCClosed():
+            subprocess.call('python3 ~/yeticold/utils/stopbitcoin.py', shell=True)
+            subprocess.call('rm -r ~/.bitcoin/yetiwarm*', shell=True)
+            subprocess.call('rm -r ~/yetiwarmwallet*', shell=True)
+            subprocess.Popen('~/yeticold/bitcoin-0.19.0rc1/bin/bitcoin-qt -proxy=127.0.0.1:9050',shell=True,start_new_session=True)
     if request.method == 'POST':
-        IBD = BTCFinished()
-        while IBD:
-            IBD = BTCFinished()
-        subprocess.call(['~/yeticold/bitcoin-0.19.0rc1/bin/bitcoin-cli createwallet "yetiwarm"'],shell=True)
-        return redirect('/YWgetseeds')
+        IBD = BTCRunning()
+        if IBD:
+            subprocess.call(['~/yeticold/bitcoin-0.19.0rc1/bin/bitcoin-cli createwallet "yeticold"'],shell=True)
+            return redirect('/YWgetseeds')
+        return redirect('/YWrestartbitcoin')
     return render_template('YWrestartbitcoin.html')
 
 #randomise priv key and get xprivs
@@ -515,20 +521,21 @@ def YWRrestartbitcoin():
     global progress
     global IBD
     global pubdesc
-    global samedesc
-    global rescan
+    global sent
     if request.method == 'GET':
-        subprocess.call('python3 ~/yeticold/utils/stopbitcoin.py', shell=True)
-        subprocess.call('rm -r ~/.bitcoin/yetiwarm*', shell=True)
-        subprocess.call('rm -r ~/yetiwarmwallet*', shell=True)
-        subprocess.Popen('~/yeticold/bitcoin-0.19.0rc1/bin/bitcoin-qt -proxy=127.0.0.1:9050',shell=True,start_new_session=True)
-        progress = BTCprogress()
+        if sent:
+            return redirect('/YWRdisplaywallet')
+        if BTCClosed():
+            subprocess.call('python3 ~/yeticold/utils/stopbitcoin.py', shell=True)
+            subprocess.call('rm -r ~/.bitcoin/yetiwarm*', shell=True)
+            subprocess.call('rm -r ~/yetiwarmwallet*', shell=True)
+            subprocess.Popen('~/yeticold/bitcoin-0.19.0rc1/bin/bitcoin-qt -proxy=127.0.0.1:9050',shell=True,start_new_session=True)
     if request.method == 'POST':
-        IBD = BTCFinished()
-        while IBD:
-            IBD = BTCFinished()
-        subprocess.call(['~/yeticold/bitcoin-0.19.0rc1/bin/bitcoin-cli createwallet "yetiwarm"'],shell=True)
-        return redirect('/YWRscandescriptor')
+        IBD = BTCRunning()
+        if IBD:
+            subprocess.call(['~/yeticold/bitcoin-0.19.0rc1/bin/bitcoin-cli createwallet "yeticold"'],shell=True)
+            return redirect('/YWRscandescriptor')
+        return redirect('/YWRrestartbitcoin')
     return render_template('YWRrestartbitcoin.html')
 
 @app.route("/YWRscandescriptor", methods=['GET', 'POST'])
@@ -625,6 +632,7 @@ def YWRdisplaywallet():
 def YWRscanrecipent():
     global error
     global receipentaddress
+    global sent
     if request.method == 'POST':
         error = None
         if request.form['option'] == 'scan':
@@ -641,6 +649,8 @@ def YWRscanrecipent():
             error = receipentaddress + ' is not a valid bitcoin address, address should have started with bc1, 3 or 1 instead of ' + receipentaddress[:1] + ', or ' + receipentaddress[:3] + '.'
         if error:
             return redirect('/YWRscanrecipent')
+        if sent:
+            return redirect('/YWRsendtransaction')
         return redirect('/YWRimportseeds')
     return render_template('YWRscanrecipent.html', error=error, recipent=receipentaddress)
 
@@ -706,6 +716,7 @@ def YWRsendtransaction():
     global receipentaddress
     global minerfee
     global transnum
+    global sent
     if request.method == 'GET':
         xpublist = pubdesc.split(',')[1:]
         xpublist[6] = xpublist[6].split('))')[0]
@@ -742,6 +753,7 @@ def YWRsendtransaction():
     if request.method == 'POST':
         response = subprocess.Popen(['~/yeticold/bitcoin-0.19.0rc1/bin/bitcoin-cli -rpcwallet=yetiwarm sendrawtransaction '+transnum['hex']+''],shell=True, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE).communicate()
         print(response)
+        sent = True
         return redirect('/YWmenu')
     return render_template('YWRsendtransaction.html', amount=amo, minerfee=minerfee, recipent=receipentaddress)
